+++ fdr-impl/src/test/java/com/dwarfeng/fdr/impl/service/PointMaintainServiceImplTest.java
@@ -12,6 +12,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
@@ -30,7 +31,11 @@ public class PointMaintainServiceImplTest {
     public void setUp() {
         points = new ArrayList<>();
         for (int i = 0; i < 5; i++) {
-            Point point = new Point(null, "name", "remark", true, true, true, true, true, true);
+            Point point = new Point(
+                    null, "name", "remark", true, true, true, true, true, true,
+                    "reservedStringAlpha", "reservedStringBravo", "reservedStringCharlie", "reservedStringDelta",
+                    12450L, 12450L, 12450, 12450, true, true, new Date(), new Date()
+            );
             points.add(point);
         }
     }