+++ fdr-sdk/src/main/java/com/dwarfeng/fdr/sdk/bean/entity/WebInputPoint.java
@@ -11,6 +11,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.groups.Default;
+import java.util.Date;
 import java.util.Objects;
 
 /**
@@ -21,7 +22,7 @@ import java.util.Objects;
  */
 public class WebInputPoint implements Bean {
 
-    private static final long serialVersionUID = 7514909111583453251L;
+    private static final long serialVersionUID = -4739858083232019144L;
 
     public static Point toStackBean(WebInputPoint webInputPoint) {
         if (Objects.isNull(webInputPoint)) {
@@ -36,7 +37,19 @@ public class WebInputPoint implements Bean {
                     webInputPoint.isFilteredKeepEnabled(),
                     webInputPoint.isFilteredPersistEnabled(),
                     webInputPoint.isTriggeredKeepEnabled(),
-                    webInputPoint.isTriggeredPersistEnabled()
+                    webInputPoint.isTriggeredPersistEnabled(),
+                    webInputPoint.getReservedStringAlpha(),
+                    webInputPoint.getReservedStringBravo(),
+                    webInputPoint.getReservedStringCharlie(),
+                    webInputPoint.getReservedStringDelta(),
+                    webInputPoint.getReservedLongAlpha(),
+                    webInputPoint.getReservedLongBravo(),
+                    webInputPoint.getReservedIntegerAlpha(),
+                    webInputPoint.getReservedIntegerBravo(),
+                    webInputPoint.getReservedBooleanAlpha(),
+                    webInputPoint.getReservedBooleanBravo(),
+                    webInputPoint.getReservedDateAlpha(),
+                    webInputPoint.getReservedDateBravo()
             );
         }
     }
@@ -73,6 +86,46 @@ public class WebInputPoint implements Bean {
 
     @JSONField(name = "triggered_persist_enabled")
     private boolean triggeredPersistEnabled;
+
+    // region 预留字段
+
+    @JSONField(name = "reserved_string_alpha")
+    private String reservedStringAlpha;
+
+    @JSONField(name = "reserved_string_bravo")
+    private String reservedStringBravo;
+
+    @JSONField(name = "reserved_string_charlie")
+    private String reservedStringCharlie;
+
+    @JSONField(name = "reserved_string_delta")
+    private String reservedStringDelta;
+
+    @JSONField(name = "reserved_long_alpha")
+    private Long reservedLongAlpha;
+
+    @JSONField(name = "reserved_long_bravo")
+    private Long reservedLongBravo;
+
+    @JSONField(name = "reserved_integer_alpha")
+    private Integer reservedIntegerAlpha;
+
+    @JSONField(name = "reserved_integer_bravo")
+    private Integer reservedIntegerBravo;
+
+    @JSONField(name = "reserved_boolean_alpha")
+    private Boolean reservedBooleanAlpha;
+
+    @JSONField(name = "reserved_boolean_bravo")
+    private Boolean reservedBooleanBravo;
+
+    @JSONField(name = "reserved_date_alpha")
+    private Date reservedDateAlpha;
+
+    @JSONField(name = "reserved_date_bravo")
+    private Date reservedDateBravo;
+
+    // endregion
 
     public WebInputPoint() {
     }
@@ -149,6 +202,103 @@ public class WebInputPoint implements Bean {
         this.triggeredPersistEnabled = triggeredPersistEnabled;
     }
 
+    public String getReservedStringAlpha() {
+        return reservedStringAlpha;
+    }
+
+    public void setReservedStringAlpha(String reservedStringAlpha) {
+        this.reservedStringAlpha = reservedStringAlpha;
+    }
+
+    public String getReservedStringBravo() {
+        return reservedStringBravo;
+    }
+
+    public void setReservedStringBravo(String reservedStringBravo) {
+        this.reservedStringBravo = reservedStringBravo;
+    }
+
+    public String getReservedStringCharlie() {
+        return reservedStringCharlie;
+    }
+
+    public void setReservedStringCharlie(String reservedStringCharlie) {
+        this.reservedStringCharlie = reservedStringCharlie;
+    }
+
+    public String getReservedStringDelta() {
+        return reservedStringDelta;
+    }
+
+    public void setReservedStringDelta(String reservedStringDelta) {
+        this.reservedStringDelta = reservedStringDelta;
+    }
+
+    public Long getReservedLongAlpha() {
+        return reservedLongAlpha;
+    }
+
+    public void setReservedLongAlpha(Long reservedLongAlpha) {
+        this.reservedLongAlpha = reservedLongAlpha;
+    }
+
+    public Long getReservedLongBravo() {
+        return reservedLongBravo;
+    }
+
+    public void setReservedLongBravo(Long reservedLongBravo) {
+        this.reservedLongBravo = reservedLongBravo;
+    }
+
+    public Integer getReservedIntegerAlpha() {
+        return reservedIntegerAlpha;
+    }
+
+    public void setReservedIntegerAlpha(Integer reservedIntegerAlpha) {
+        this.reservedIntegerAlpha = reservedIntegerAlpha;
+    }
+
+    public Integer getReservedIntegerBravo() {
+        return reservedIntegerBravo;
+    }
+
+    public void setReservedIntegerBravo(Integer reservedIntegerBravo) {
+        this.reservedIntegerBravo = reservedIntegerBravo;
+    }
+
+    public Boolean getReservedBooleanAlpha() {
+        return reservedBooleanAlpha;
+    }
+
+    public void setReservedBooleanAlpha(Boolean reservedBooleanAlpha) {
+        this.reservedBooleanAlpha = reservedBooleanAlpha;
+    }
+
+    public Boolean getReservedBooleanBravo() {
+        return reservedBooleanBravo;
+    }
+
+    public void setReservedBooleanBravo(Boolean reservedBooleanBravo) {
+        this.reservedBooleanBravo = reservedBooleanBravo;
+    }
+
+    public Date getReservedDateAlpha() {
+        return reservedDateAlpha;
+    }
+
+    public void setReservedDateAlpha(Date reservedDateAlpha) {
+        this.reservedDateAlpha = reservedDateAlpha;
+    }
+
+    public Date getReservedDateBravo() {
+        return reservedDateBravo;
+    }
+
+    public void setReservedDateBravo(Date reservedDateBravo) {
+        this.reservedDateBravo = reservedDateBravo;
+    }
+
+    @SuppressWarnings("DuplicatedCode")
     @Override
     public String toString() {
         return "WebInputPoint{" +
@@ -161,6 +311,18 @@ public class WebInputPoint implements Bean {
                 ", filteredPersistEnabled=" + filteredPersistEnabled +
                 ", triggeredKeepEnabled=" + triggeredKeepEnabled +
                 ", triggeredPersistEnabled=" + triggeredPersistEnabled +
+                ", reservedStringAlpha='" + reservedStringAlpha + '\'' +
+                ", reservedStringBravo='" + reservedStringBravo + '\'' +
+                ", reservedStringCharlie='" + reservedStringCharlie + '\'' +
+                ", reservedStringDelta='" + reservedStringDelta + '\'' +
+                ", reservedLongAlpha=" + reservedLongAlpha +
+                ", reservedLongBravo=" + reservedLongBravo +
+                ", reservedIntegerAlpha=" + reservedIntegerAlpha +
+                ", reservedIntegerBravo=" + reservedIntegerBravo +
+                ", reservedBooleanAlpha=" + reservedBooleanAlpha +
+                ", reservedBooleanBravo=" + reservedBooleanBravo +
+                ", reservedDateAlpha=" + reservedDateAlpha +
+                ", reservedDateBravo=" + reservedDateBravo +
                 '}';
     }
 }